Initialize Web Chat in the Root component
+ Instead of importing it as a plugin, we can drop the embed
  script in the root component (which never unmounts)
